WIP - Refactor Initiative component selectors and imports
