+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 import { Provider } from 'react-redux';
-import {store} from './store/store.js'
+import { store } from './store/store'
 import "babel-polyfill";
 
 ReactDOM.render(
