+++ src/services/api/auth/index.js
@@ -1,3 +1,4 @@
 import { userReducer } from './duck';
+import { * as selectors } from './selectors';
 
-export { userReducer };
+export { userReducer, selectors };
